Update to v1.8.51
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@138 01821946-851b-0410-b7f4-f1477c70e57a

+++ main/physics.h
@@ -77,7 +77,7 @@ COPYRIGHT 1993-1999 PARALLAX SOFTWARE CORPORATION.  ALL RIGHTS RESERVED.
 
 // these global vars are set after a call to DoPhysicsSim().  Ugly, I know.
 // list of segments went through
-extern int physSegList [MAX_FVI_SEGS], nPhysSegs;
+extern short physSegList [MAX_FVI_SEGS], nPhysSegs;
 
 // Read contrls and set physics vars
 void ReadFlyingControls(tObject *obj);
